refactor (#6) : Domain 간 디렉토리 분리 (역할 분리)
