--- api/src/main/java/com/yapp/api/domain/album/comment/persistence/handler/CommentQueryHandler.java
@@ -1,3 +1,0 @@
-package com.yapp.api.domain.album.comment.persistence.handler;
-
-public interface CommentQueryHandler {}
