--- api/src/main/java/com/yapp/api/domain/album/comment/persistence/repository/CommentRepository.java
@@ -1,7 +1,0 @@
-package com.yapp.api.domain.album.comment.persistence.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.yapp.api.domain.album.comment.persistence.entity.Comment;
-
-public interface CommentRepository extends JpaRepository<Comment, Long> {}
