+++ api/src/main/java/com/yapp/api/domain/album/controller/AlbumCommandApi.java
@@ -2,8 +2,6 @@ package com.yapp.api.domain.album.controller;
 
 import static org.springframework.http.MediaType.*;
 
-import java.util.List;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -11,9 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
 
 import com.yapp.api.domain.album.controller.dto.AlbumRequest;
 import com.yapp.api.domain.album.controller.dto.AlbumResponse;
@@ -23,11 +19,6 @@ import lombok.RequiredArgsConstructor;
 @RestController
 @RequiredArgsConstructor
 public class AlbumCommandApi {
-	@PostMapping(value = "/album/files", consumes = MULTIPART_FORM_DATA_VALUE, produces = APPLICATION_JSON_VALUE)
-	ResponseEntity<AlbumResponse.UploadFiles> uploadFiles(@RequestPart List<MultipartFile> files) {
-		return null;
-	}
-
 	@PostMapping(value = "/album/photos", consumes = APPLICATION_JSON_VALUE)
 	ResponseEntity<Void> uploadPhotos(@RequestBody AlbumRequest.Upload request) {
 		return null;
@@ -51,11 +42,6 @@ public class AlbumCommandApi {
 
 	@DeleteMapping(value = "/album/{albumId}")
 	ResponseEntity<Void> removeAlbum(@PathVariable(value = "albumId") Long albumId) {
-		return null;
-	}
-
-	@DeleteMapping(value = "/album/{fileId}")
-	ResponseEntity<Void> removeFile(@PathVariable(value = "fileId") Long fileId) {
 		return null;
 	}
