+++ api/src/main/java/com/yapp/api/domain/album/element/comment/persistence/handler/CommentCommandHandler.java
@@ -1,0 +1,3 @@
+package com.yapp.api.domain.album.element.comment.persistence.handler;
+
+public interface CommentCommandHandler {}
