+++ api/src/main/java/com/yapp/api/domain/album/element/folder/persistence/handler/FolderQueryHandler.java
@@ -1,0 +1,3 @@
+package com.yapp.api.domain.album.element.folder.persistence.handler;
+
+public interface FolderQueryHandler {}
