--- api/src/main/java/com/yapp/api/domain/album/file/persistence/handler/FileCommandHandler.java
@@ -1,3 +1,0 @@
-package com.yapp.api.domain.album.file.persistence.handler;
-
-public interface FileCommandHandler {}
