--- api/src/main/java/com/yapp/api/domain/album/file/persistence/handler/FileQueryHandler.java
@@ -1,3 +1,0 @@
-package com.yapp.api.domain.album.file.persistence.handler;
-
-public class FileQueryHandler {}
