--- api/src/main/java/com/yapp/api/domain/album/folder/persistence/handler/FolderCommandHandler.java
@@ -1,3 +1,0 @@
-package com.yapp.api.domain.album.folder.persistence.handler;
-
-public interface FolderCommandHandler {}
