--- api/src/main/java/com/yapp/api/domain/calendar/appointment/persistence/handler/AppointmentCommandHandler.java
@@ -1,3 +1,0 @@
-package com.yapp.api.domain.calendar.appointment.persistence.handler;
-
-public interface AppointmentCommandHandler {}
