--- api/src/main/java/com/yapp/api/domain/calendar/appointment/persistence/handler/AppointmentQueryHandler.java
@@ -1,5 +1,0 @@
-package com.yapp.api.domain.calendar.appointment.persistence.handler;
-
-public interface AppointmentQueryHandler {
-
-}
