--- api/src/main/java/com/yapp/api/domain/calendar/appointment/persistence/repository/AppointmentRepository.java
@@ -1,7 +1,0 @@
-package com.yapp.api.domain.calendar.appointment.persistence.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.yapp.api.domain.calendar.appointment.persistence.entity.Appointment;
-
-public interface AppointmentRepository extends JpaRepository<Appointment, Long> {}
