--- api/src/main/java/com/yapp/api/domain/calendar/appointment/service/AppointmentService.java
@@ -1,17 +1,0 @@
-package com.yapp.api.domain.calendar.appointment.service;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import com.yapp.api.domain.calendar.appointment.persistence.handler.AppointmentCommandHandler;
-import com.yapp.api.domain.calendar.appointment.persistence.handler.AppointmentQueryHandler;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class AppointmentService {
-	private final AppointmentCommandHandler appointmentCommandHandler;
-	private final AppointmentQueryHandler appointmentQueryHandler;
-}
