+++ api/src/main/java/com/yapp/api/domain/calendar/element/appointment/persistence/handler/AppointmentCommandHandler.java
@@ -1,0 +1,3 @@
+package com.yapp.api.domain.calendar.element.appointment.persistence.handler;
+
+public interface AppointmentCommandHandler {}
