+++ api/src/main/java/com/yapp/api/domain/calendar/element/appointment/persistence/handler/AppointmentQueryHandler.java
@@ -1,0 +1,5 @@
+package com.yapp.api.domain.calendar.element.appointment.persistence.handler;
+
+public interface AppointmentQueryHandler {
+
+}
