+++ api/src/main/java/com/yapp/api/domain/calendar/element/appointment/persistence/repository/AppointmentRepository.java
@@ -1,0 +1,7 @@
+package com.yapp.api.domain.calendar.element.appointment.persistence.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.yapp.api.domain.calendar.element.appointment.persistence.entity.Appointment;
+
+public interface AppointmentRepository extends JpaRepository<Appointment, Long> {}
