+++ api/src/main/java/com/yapp/api/domain/calendar/element/visibility/persistence/handler/VisibilityCommandHandler.java
@@ -1,0 +1,3 @@
+package com.yapp.api.domain.calendar.element.visibility.persistence.handler;
+
+public interface VisibilityCommandHandler {}
