+++ api/src/main/java/com/yapp/api/domain/calendar/element/visibility/persistence/repository/VisibilityRepository.java
@@ -1,0 +1,7 @@
+package com.yapp.api.domain.calendar.element.visibility.persistence.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.yapp.api.domain.calendar.element.visibility.persistence.entity.Visibility;
+
+public interface VisibilityRepository extends JpaRepository<Visibility, Long> {}
