--- api/src/main/java/com/yapp/api/domain/calendar/visibility/persistence/handler/VisibilityCommandHandler.java
@@ -1,3 +1,0 @@
-package com.yapp.api.domain.calendar.visibility.persistence.handler;
-
-public interface VisibilityCommandHandler {}
