--- api/src/main/java/com/yapp/api/domain/calendar/visibility/persistence/handler/VisibilityQueryHandler.java
@@ -1,3 +1,0 @@
-package com.yapp.api.domain.calendar.visibility.persistence.handler;
-
-public interface VisibilityQueryHandler {}
