+++ api/src/main/java/com/yapp/api/domain/file/persistence/handler/FileCommandHandler.java
@@ -1,0 +1,3 @@
+package com.yapp.api.domain.file.persistence.handler;
+
+public interface FileCommandHandler {}
