+++ api/src/main/java/com/yapp/api/domain/file/persistence/handler/FileQueryHandler.java
@@ -1,0 +1,3 @@
+package com.yapp.api.domain.file.persistence.handler;
+
+public class FileQueryHandler {}
