+++ api/src/main/java/com/yapp/api/domain/user/persistence/entity/User.java
@@ -8,7 +8,6 @@ import java.util.Set;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
